Replace execute_tx_configurable_with_state with execute_tx_with_blockifier
