+++ replay/src/benchmark.rs
@@ -1,11 +1,8 @@
 use crate::{build_cached_state, get_transaction_hashes};
-use blockifier::{
-    blockifier::block::BlockInfo,
-    context::{BlockContext, ChainInfo},
-    state::cached_state::CachedState,
-    versioned_constants::VersionedConstants,
+use blockifier::{context::BlockContext, state::cached_state::CachedState};
+use rpc_state_reader::blockifier_state_reader::{
+    execute_tx_with_blockifier, fetch_block_context, RpcStateReader,
 };
-use rpc_state_reader::blockifier_state_reader::{execute_tx_with_blockifier, RpcStateReader};
 use starknet_api::{
     block::BlockNumber,
     hash::StarkFelt,
@@ -38,13 +35,13 @@ pub fn fetch_block_range_data(
         let block_number = BlockNumber(block_number);
 
         // Create a cached state
-        let state = build_cached_state(&chain, block_number.0 - 1);
+        let state = build_cached_state(chain, block_number.0 - 1);
 
         // Fetch block context
         let block_context = fetch_block_context(&state, block_number);
 
         // Fetch transactions for the block
-        let transactions = get_transaction_hashes(&chain, block_number.0)
+        let transactions = get_transaction_hashes(chain, block_number.0)
             .unwrap()
             .into_iter()
             .map(|transaction_hash| {
@@ -63,29 +60,6 @@ pub fn fetch_block_range_data(
     }
 
     block_caches
-}
-
-fn fetch_block_context(
-    state: &CachedState<RpcStateReader>,
-    block_number: BlockNumber,
-) -> BlockContext {
-    let rpc_block_info = state.state.0.get_block_info().unwrap();
-    let gas_price = state.state.0.get_gas_price(block_number.0).unwrap();
-
-    BlockContext::new_unchecked(
-        &BlockInfo {
-            block_number,
-            block_timestamp: rpc_block_info.block_timestamp,
-            sequencer_address: rpc_block_info.sequencer_address,
-            gas_prices: gas_price,
-            use_kzg_da: false,
-        },
-        &ChainInfo {
-            chain_id: state.state.0.get_chain_name().into(),
-            fee_token_addresses: Default::default(),
-        },
-        &VersionedConstants::latest_constants_with_overrides(u32::MAX, usize::MAX),
-    )
 }
 
 /// Executes the given block range, discarding any state changes applied to it
